Aarch64 support
Add support for Aarch64 (arm64). A few things are still to be done or
do not work (FPU handling, error handling when incorrectly returning
from a co without co_exit()) but the basic examples at least do work.

Some changes are not strictly necessary such as changing a few macros
to static inline functions. I did them because the code looks cleaner
this way and it helped with debugging.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>